requirements.txt
streamlit
google-generativeai
python-dotenv
fpdf
python-dotenv
tqdm